Move actions from simulation to a separate file
